compiler: Find columns in default schemas

+++ internal/endtoend/testdata/select_star/sqlite/go/models.go
@@ -7,8 +7,8 @@ import (
 )
 
 type User struct {
-	ID        int32
+	ID        int64
 	FirstName string
 	LastName  sql.NullString
-	Age       int32
+	Age       int64
 }
